% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_graphs.R
\name{Check_graphs}
\alias{Check_graphs}
\title{Check Graphs}
\usage{
Check_graphs(
  Name_File = "RATIO",
  Name_Col_Outcome = "Outcome",
  Outcome_of_interest = "soil carbon",
  Intervention = "Sub_Cat_intervention",
  Effect_size = "Effect size"
)
}
\arguments{
\item{name_file}{Name of the file to use as input data. Default: 'RATIO'.}

\item{name_col_outcome}{Name of the column in the input data that contains the outcome variable. Default: 'Outcome'.}

\item{outcome_of_interest}{Outcome variable of interest. Default: 'soil carbon'.}

\item{intervention}{Name of the column in the input data that contains the intervention variable. Default: 'Sub_Cat_intervention'.}

\item{effect_size}{Name of the column in the input data that contains the effect size variable. Default: 'Effect size'.}
}
\value{
A list containing the variance plots and the forest plot.
}
\description{
Creates a forest plot and two variance plots to check data.
}
\details{
This function takes input data and creates a forest plot and two variance plots to check the data.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{facet_wrap}}, \code{\link[ggplot2]{geom_abline}}, \code{\link[ggplot2]{geom_crossbar}}, \code{\link[ggplot2]{theme}}, \code{\link[ggplot2]{scale_size}}
\code{\link[ggpubr]{ggpubr}}, \code{\link[ggpubr]{theme_pubr}}
\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}
}
