% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_Effect_sizes.R
\name{FUN_CI_from_var}
\alias{FUN_CI_from_var}
\title{Calculate confidence interval from variance}
\usage{
FUN_CI_from_var(Var)
}
\arguments{
\item{Var}{a numeric value representing the variance}
}
\value{
a numeric value representing the confidence interval
}
\description{
This function calculates the confidence interval given the variance using the formula CI = sqrt(Var)/21.96
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
FUN_CI_from_var(4)
}
}
}
