% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_Effect_sizes.R
\name{FUN_d_to_g}
\alias{FUN_d_to_g}
\title{Convertir d en Hedges' g}
\usage{
FUN_d_to_g(d, df)
}
\arguments{
\item{d}{La taille de l'effet d à convertir}

\item{df}{Les degrés de liberté associés à d}
}
\value{
La taille de l'effet Hedges' g correspondante
}
\description{
Convertir d en Hedges' g
}
\details{
Convertit la taille de l'effet d (Cohen's d) en taille de l'effet Hedges' g en utilisant le facteur de correction J.
J est une approximation couramment utilisée pour convertir d en g.
La formule exacte de J a été décrite par Hedges (1981).
}
\examples{
\dontrun{
if(interactive()){
#EXAMPLE1
}
}
}
\references{
Hedges, L. V. (1981). Distribution theory for Glass's estimator of effect size and related estimators. Journal of Educational Statistics, 6(2), 107-128.
}
\keyword{Cohen's}
\keyword{Hedges'}
\keyword{d,}
\keyword{de}
\keyword{g,}
\keyword{l'effet}
\keyword{taille}
