% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_Effect_sizes.R
\name{FUN_LOR_to_d}
\alias{FUN_LOR_to_d}
\title{Convertir d'un odds ratio en Cohen's d}
\usage{
FUN_LOR_to_d(LOR)
}
\arguments{
\item{LOR}{Le odds ratio à convertir.}
}
\value{
La valeur de Cohen's d.
}
\description{
Cette fonction permet de convertir un odds ratio en Cohen's d.
}
\details{
La formule utilisée pour convertir un odds ratio en Cohen's d est : d = LOR * sqrt(3) / pi.
}
\examples{
\dontrun{
if(interactive()){
# Exemple d'utilisation :
FUN_LOR_to_d(1.5)
# Le résultat attendu est : 0.2461187
}
}
}
