% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_MAT_RED.R
\name{FUN_MAT_RED}
\alias{FUN_MAT_RED}
\title{Calculate the pseudo-variance-covariance matrix based on variance and redundancy}
\usage{
FUN_MAT_RED(
  Matrix_name = "Global_Redundancy_matrix.xlsx",
  Selected_rows_ID,
  Variance
)
}
\arguments{
\item{Variance}{Variance of the effect-sizes of the selected meta-analyses.}

\item{Global_Redundancy_matrix}{Global matrix of redudancy of primary studies between meta-analyses calculated with the Global_Redundancy_matrix function, Default: 'Matrice_red.csv'}

\item{Selected_ID}{Selected ID of the meta-analyses (E.g. for one type of intervention)}
}
\value{
a pseudo variance co-variance matrix to be used in the final meta-analytical model
}
\description{
Based on DOI and al., 2008
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{fread}}
\code{\link[here]{here}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{arrange}}
\code{\link[base]{subset}}, \code{\link[base]{merge}}, \code{\link[base]{replace}}, \code{\link[base]{length}}, \code{\link[base]{split}}
\code{\link[tidyr]{gather}}
}
