% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_Effect_sizes.R
\name{FUN_r_to_d}
\alias{FUN_r_to_d}
\title{Convert correlation coefficient to standardized mean difference (Cohen's d)}
\usage{
FUN_r_to_d(r)
}
\arguments{
\item{r}{a numeric value representing a correlation coefficient}
}
\value{
a numeric value representing the corresponding standardized mean difference (Cohen's d)
}
\description{
This function converts a correlation coefficient to Cohen's d using the formula d = 2r/(sqrt(1-r^2))
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
FUN_r_to_d(0.5)
}
}
}
