% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_Effect_sizes.R
\name{FUN_Var_from_CI}
\alias{FUN_Var_from_CI}
\title{Calculate variance from confidence interval}
\usage{
FUN_Var_from_CI(CI)
}
\arguments{
\item{CI}{A numeric vector representing the confidence interval.}
}
\value{
The variance calculated from the confidence interval.
}
\description{
This function calculates the variance from a given confidence interval.
}
\details{
The formula used to calculate the variance is Var = (CI / 1.96 * 2) ^ 2, where 1.96 is the z-score for a 95\% confidence interval.
}
\examples{
\dontrun{
if(interactive()){
#EXAMPLE1
CI <- c(2.5, 5.5)
FUN_Var_from_CI(CI)
}
}
}
