% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_Effect_sizes.R
\name{FUN_var_LOR_to_d}
\alias{FUN_var_LOR_to_d}
\title{Convert variance of log odds ratio to variance of Cohen's d}
\usage{
FUN_var_LOR_to_d(Var_LOR)
}
\arguments{
\item{Var_LOR}{Variance of log odds ratio}
}
\value{
Variance of Cohen's d
}
\description{
This function converts the variance of log odds ratio (Var_LOR) to the variance of Cohen's d (Var_d).
}
\details{
This function is used in meta-analyses when combining studies that report log odds ratios and studies that report Cohen's d.
}
\examples{
\dontrun{
if(interactive()){
  # EXAMPLE 1: Convert variance of log odds ratio to variance of Cohen's d
  Var_LOR <- 0.01
  Var_d <- FUN_var_LOR_to_d(Var_LOR)
  print(Var_d)
}
}
}
