% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_Effect_sizes.R
\name{FUN_Z_to_r}
\alias{FUN_Z_to_r}
\title{Convert Z score to correlation coefficient}
\usage{
FUN_Z_to_r(Z)
}
\arguments{
\item{Z}{a numeric value representing a Z score}
}
\value{
a numeric value representing the corresponding correlation coefficient
}
\description{
This function converts a Z score to a correlation coefficient using the formula r = (exp(2Z)-1)/(exp(2Z)+1)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
FUN_Z_to_r(2)
}
}
}
