% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Redundancy_matrix.R
\name{Global_Redundancy_matrix}
\alias{Global_Redundancy_matrix}
\title{FUNCTION_TITLE}
\usage{
Global_Redundancy_matrix(
  File_name = "Data_Base_C_Sol_2023-02-18.xlsx",
  sheet_name = "Primary_studies",
  ID = "ID",
  DOI = "DOI",
  FACTOR = 1
)
}
\arguments{
\item{sheet_name}{Name of the sheet where the primary studies are.}

\item{ID}{Names of the column for the ID of the meta-analyses}

\item{DOI}{Names of the column for the identifier of the primary studies}

\item{FACTOR}{Factor for the Meta-analyses with no available Primary studies.Between 0 and 1. 0-> MA with no studies = no redundance with the others, 1-> MA with no studies= 100\% redundant with the others}

\item{file_name}{Names of the file (Excel file)}
}
\value{
return a matrix of redudancy between all Meta-analyses (in count, not percent)

return a matrix of redudancy between all Meta-analyses (in count, not percent)
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
\code{\link[here]{here}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{filter}}
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_bar}}, \code{\link[ggplot2]{coord_flip}}, \code{\link[ggplot2]{labs}}
\code{\link[data.table]{dcast.data.table}}, \code{\link[data.table]{fwrite}}
}
