% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Data2023.R
\name{Load_Data}
\alias{Load_Data}
\title{Load data from Excel file}
\usage{
Load_Data(
  File_name = "Data_Base_C_Sol_2023-02-18.xlsx",
  sheet_PS = "Primary_studies",
  sheet_desMA = "retained_meta-analyses",
  sheet_ES = "Effect-sizes"
)
}
\arguments{
\item{File_name}{Name of the Excel file to load data from. Default: 'Data_Base_C_Sol_2023-02-18.xlsx'}

\item{sheet_PS}{Name of the sheet in the Excel file containing primary studies data. Default: 'Primary_studies'}

\item{sheet_desMA}{Name of the sheet in the Excel file containing retained meta-analyses data. Default: 'retained_meta-analyses'}

\item{sheet_ES}{Name of the sheet in the Excel file containing effect sizes data. Default: 'Effect-sizes'}
}
\value{
A list containing three data frames: primary studies, retained meta-analyses, and effect sizes.
}
\description{
This function loads data from an Excel file and returns three data frames: primary studies, retained meta-analyses, and effect sizes.
}
\details{
The function reads data from an Excel file, formats it, and returns three data frames: primary studies, retained meta-analyses, and effect sizes.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
\code{\link[here]{here}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr]{group_by}}
\code{\link[forcats]{fct_recode}}
}
