% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_DOI.R
\name{QUALITY_DOI}
\alias{QUALITY_DOI}
\title{Integrate the quality score into the variance}
\usage{
QUALITY_DOI(SCORE, vi)
}
\arguments{
\item{SCORE}{Quality score of the meta-analysis (numeric)}

\item{vi}{Variance of the effect-size}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
BAsed on Doiu et al., 2008.https://www.jstor.org/stable/20486499
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
